Backport: extracting unsafe proxy code in an Unsafe class - no fonctional change
