+++ webbeans-impl/src/main/java/org/apache/webbeans/proxy/AbstractProxyFactory.java
@@ -18,31 +18,24 @@
  */
 package org.apache.webbeans.proxy;
 
-import static org.apache.xbean.asm6.ClassReader.SKIP_DEBUG;
-
-import java.io.InputStream;
-import java.lang.reflect.Constructor;
-import java.lang.reflect.Field;
-import java.lang.reflect.InvocationTargetException;
-import java.lang.reflect.Method;
-import java.lang.reflect.Modifier;
-import java.security.AccessController;
-import java.security.PrivilegedAction;
-import java.security.ProtectionDomain;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.logging.Logger;
-
 import org.apache.webbeans.config.WebBeansContext;
 import org.apache.webbeans.exception.ProxyGenerationException;
 import org.apache.webbeans.exception.WebBeansException;
-import org.apache.webbeans.logger.WebBeansLoggerFacade;
 import org.apache.xbean.asm6.ClassReader;
 import org.apache.xbean.asm6.ClassWriter;
 import org.apache.xbean.asm6.MethodVisitor;
 import org.apache.xbean.asm6.Opcodes;
 import org.apache.xbean.asm6.Type;
 import org.apache.xbean.asm6.shade.commons.EmptyVisitor;
+
+import java.io.InputStream;
+import java.lang.reflect.Constructor;
+import java.lang.reflect.Method;
+import java.lang.reflect.Modifier;
+import java.util.ArrayList;
+import java.util.Arrays;
+
+import static org.apache.xbean.asm6.ClassReader.SKIP_DEBUG;
 
 /**
  * Base class for all OWB Proxy factories
@@ -58,20 +51,9 @@ public abstract class AbstractProxyFactory
      */
     public static final int MODIFIER_VARARGS = 0x00000080;;
 
-
-    private static final Logger logger = WebBeansLoggerFacade.getLogger(AbstractProxyFactory.class);
-
+    protected final Unsafe unsafe;
 
     protected WebBeansContext webBeansContext;
-
-    /**
-     * contains the instance of sun.misc.Unsafe.
-     * We use it for creating the proxy instance without fully
-     * initializing the class.
-     */
-    private Object unsafe = null;
-    private Method unsafeAllocateInstance = null;
-    private Method unsafeDefineClass;
 
     private final int javaVersion;
 
@@ -88,7 +70,7 @@ public abstract class AbstractProxyFactory
     {
         this.webBeansContext = webBeansContext;
         javaVersion = determineDefaultJavaVersion();
-        initializeUnsafe();
+        unsafe = new Unsafe();
     }
 
     private int determineDefaultJavaVersion()
@@ -250,7 +232,7 @@ public abstract class AbstractProxyFactory
                 sortOutDuplicateMethods(nonInterceptedMethods),
                 constructor);
 
-        return defineAndLoadClass(classLoader, proxyClassName, proxyBytes);
+        return unsafe.defineAndLoadClass(classLoader, proxyClassName, proxyBytes);
     }
 
     private Method[] sortOutDuplicateMethods(Method[] methods)
@@ -354,79 +336,6 @@ public abstract class AbstractProxyFactory
         }
         // mainly for JVM classes - outside the classloader, find to fallback on the JVM version
         return javaVersion;
-    }
-
-
-    /**
-     * The 'defineClass' method on the ClassLoader is protected, thus we need to invoke it via reflection.
-     * @return the Class which got loaded in the classloader
-     */
-    private <T> Class<T> defineAndLoadClass(ClassLoader classLoader, String proxyName, byte[] proxyBytes)
-            throws ProxyGenerationException
-    {
-        Class<?> clazz = classLoader.getClass();
-
-        Method defineClassMethod = null;
-        do
-        {
-            try
-            {
-                defineClassMethod = clazz.getDeclaredMethod("defineClass", String.class, byte[].class, int.class, int.class);
-            }
-            catch (NoSuchMethodException e)
-            {
-                // do nothing, we need to search the superclass
-            }
-
-            clazz = clazz.getSuperclass();
-        } while (defineClassMethod == null && clazz != Object.class);
-
-        if (defineClassMethod != null && !defineClassMethod.isAccessible())
-        {
-            try
-            {
-                defineClassMethod.setAccessible(true);
-            }
-            catch (RuntimeException re) // likely j9, let's use unsafe
-            {
-                defineClassMethod = null;
-            }
-        }
-
-        try
-        {
-            final Class<T> definedClass;
-
-            if (defineClassMethod != null)
-            {
-                definedClass = (Class<T>) defineClassMethod.invoke(classLoader, proxyName, proxyBytes, 0, proxyBytes.length);
-            }
-            else
-            {
-                definedClass = (Class<T>) unsafeDefineClass.invoke(unsafe, proxyName, proxyBytes, 0, proxyBytes.length, classLoader, null);
-            }
-
-            return (Class<T>) Class.forName(definedClass.getName(), true, classLoader);
-        }
-        catch (InvocationTargetException le) // if concurrent calls are done then ensure to just reload the created one
-        {
-            if (LinkageError.class.isInstance(le.getCause()))
-            {
-                try
-                {
-                    return (Class<T>) Class.forName(proxyName.replace('/', '.'), true, classLoader);
-                }
-                catch (ClassNotFoundException e)
-                {
-                    // default error handling
-                }
-            }
-            throw new ProxyGenerationException(le.getCause());
-        }
-        catch (Throwable e)
-        {
-            throw new ProxyGenerationException(e);
-        }
     }
 
 
@@ -689,131 +598,6 @@ public abstract class AbstractProxyFactory
     {
         final Class<?> returnType = delegatedMethod.getReturnType();
         mv.visitInsn(getReturnInsn(returnType));
-    }
-
-    protected <T> T unsafeNewInstance(Class<T> clazz)
-    {
-        try
-        {
-            if (unsafeAllocateInstance != null)
-            {
-                return (T) unsafeAllocateInstance.invoke(unsafe, clazz);
-            }
-            else
-            {
-                try
-                {
-                    return (T) clazz.newInstance();
-                }
-                catch (InstantiationException e)
-                {
-                    throw new IllegalStateException("Failed to allocateInstance of Proxy class " + clazz.getName(), e);
-                }
-            }
-        }
-        catch (IllegalAccessException e)
-        {
-            throw new IllegalStateException("Failed to allocateInstance of Proxy class " + clazz.getName(), e);
-        }
-        catch (InvocationTargetException e)
-        {
-            Throwable throwable = e.getTargetException() != null ? e.getTargetException() : e;
-            throw new IllegalStateException("Failed to allocateInstance of Proxy class " + clazz.getName(),
-                    throwable);
-        }
-    }
-
-
-    private void initializeUnsafe()
-    {
-        final Class<?> unsafeClass;
-        try
-        {
-            unsafeClass = AccessController.doPrivileged(new PrivilegedAction<Class<?>>()
-            {
-                @Override
-                public Class<?> run()
-                {
-                    try
-                    {
-                        return Thread.currentThread().getContextClassLoader().loadClass("sun.misc.Unsafe");
-                    }
-                    catch (Exception e)
-                    {
-                        try
-                        {
-                            return ClassLoader.getSystemClassLoader().loadClass("sun.misc.Unsafe");
-                        }
-                        catch (ClassNotFoundException e1)
-                        {
-                            throw new IllegalStateException("Cannot get sun.misc.Unsafe", e);
-                        }
-                    }
-                }
-            });
-        }
-        catch (Exception e)
-        {
-            throw new IllegalStateException("Cannot get sun.misc.Unsafe class", e);
-        }
-
-        Object unsafe = AccessController.doPrivileged(new PrivilegedAction<Object>()
-        {
-            @Override
-            public Object run()
-            {
-                try
-                {
-                    Field field = unsafeClass.getDeclaredField("theUnsafe");
-                    field.setAccessible(true);
-                    return field.get(null);
-                }
-                catch (Exception e)
-                {
-                    logger.warning("ATTENTION: Cannot get sun.misc.Unsafe - will use newInstance() instead! Intended for GAE only!");
-                    return null;
-                }
-            }
-        });
-
-        this.unsafe = unsafe;
-
-        if (unsafe != null)
-        {
-            unsafeAllocateInstance = AccessController.doPrivileged(new PrivilegedAction<Method>()
-            {
-                @Override
-                public Method run()
-                {
-                    try
-                    {
-                        Method mtd = unsafeClass.getDeclaredMethod("allocateInstance", Class.class);
-                        mtd.setAccessible(true);
-                        return mtd;
-                    }
-                    catch (Exception e)
-                    {
-                        throw new IllegalStateException("Cannot get sun.misc.Unsafe.allocateInstance", e);
-                    }
-                }
-            });
-            unsafeDefineClass = AccessController.doPrivileged(new PrivilegedAction<Method>()
-            {
-                @Override
-                public Method run()
-                {
-                    try
-                    {
-                        return unsafeClass.getDeclaredMethod("defineClass",
-                                String.class, byte[].class, int.class, int.class, ClassLoader.class, ProtectionDomain.class);
-                    }
-                    catch (Exception e)
-                    {
-                        throw new IllegalStateException("Cannot get Unsafe.defineClass", e);
-                    }
-                }
-            });
-        }
     }
 
     /**
